AI/dyamic_classification_model/requirements.txt
torch==2.6.0
torchvision
torchaudio
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git
pytorch-lightning
fvcore
iopath
av
decord
pyyaml